requirements.txt
langchain-core
langchain>=0.2.0
langchain-text-splitters>=0.2.0
langchain-google-genai>=1.0.0
langchain-community>=0.2.0
langchain-huggingface>=0.0.3
langchain-chroma>=0.1.0
chromadb>=0.5.0
google-generativeai>=0.7.0
sentence-transformers
python-dotenv
